LayoutTests/imported/w3c/web-platform-tests/css/css-multicol/multicol-fill-balance-014-expected.txt

Test passes if there is a filled green square below.


FAIL multicol-fill-balance-014 assert_equals: expected 100 but got 70
